move indexing impl to new mod
